--- a/.github/workflows/release-pypi.yml
+++ b/.github/workflows/release-pypi.yml
@@ -15,24 +15,17 @@ jobs:
       run:
         working-directory: python
     steps:
-      - name: Checkout code
-        uses: actions/checkout@v5
+      - uses: actions/checkout@v5
 
-      - name: Set up Python 3.13
-        uses: actions/setup-python@v5
+      - name: Install uv and set the python version
+        uses: astral-sh/setup-uv@v6
         with:
+          enable-cache: true
           python-version: "3.13"
-
-      - name: Show Python version
-        run: python --version
+          version: "0.8.17"
 
       - name: Build package
-        run: |
-          python -m pip install build twine
-          python -m build
-          twine check --strict dist/*
+        run: uv build
 
-      - name: Publish package distributions to PyPI
-        uses: pypa/gh-action-pypi-publish@release/v1
-        with:
-          packages-dir: python/dist
+      - name: Publish package
+        run: uv publish