--- a/.github/workflows/release-pypi.yml
+++ b/.github/workflows/release-pypi.yml
@@ -28,4 +28,4 @@ jobs:
         run: uv build
 
       - name: Publish package
-        run: uv publish
+        run: uv publish --trusted-publishing=always