--- a/.github/workflows/test-python.yml
+++ b/.github/workflows/test-python.yml
@@ -41,7 +41,6 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -U setuptools
           pip install tox tox-gh-actions codecov
       - name: Test with tox
         working-directory: ./python