--- a/.github/workflows/test-python.yml
+++ b/.github/workflows/test-python.yml
@@ -21,32 +21,33 @@ on:
 
 jobs:
   build:
-
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
         include:
-          # Test latest python on Windows / macOS
+          # Test latest python on Windows and macOS
           - { os: 'windows-latest', python-version: '3.13' }
           - { os: 'macos-latest', python-version: '3.13' }
         os: ['ubuntu-latest']
         python-version: ['3.9', '3.10', '3.11', '3.12', '3.13', 'pypy3.9', 'pypy3.10']
-
+    defaults:
+      run:
+        working-directory: python
     steps:
       - uses: actions/checkout@v5
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v6
+
+      - name: Install uv and set the python version
+        uses: astral-sh/setup-uv@v6
         with:
+          enable-cache: true
           python-version: ${{ matrix.python-version }}
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install tox tox-gh-actions codecov
-      - name: Test with tox
-        working-directory: ./python
-        run: |
-          tox
+          version: "0.8.17"
+
+      - name: Linting and testing
+        run: uv run tox -e="lint,${{ matrix.python-version }}"
+
       - name: Gather codecov report
-        working-directory: ./python
-        run: |
-          codecov
+        run: uvx codecov
+
+      - name: Build package
+        run: uv build