--- a/.github/workflows/test-python.yml
+++ b/.github/workflows/test-python.yml
@@ -7,6 +7,7 @@ on:
       - main
       - renovate/**
     paths:
+      - .pre-commit-config.yaml
       - python/**
       - Makefile
       - .github/**
@@ -14,6 +15,7 @@ on:
     branches:
       - main
     paths:
+      - .pre-commit-config.yaml
       - python/**
       - Makefile
       - .github/**