--- a/python/.gitignore
+++ b/python/.gitignore
@@ -53,3 +53,6 @@ nosetests.xml
 /include
 /share
 /local
+
+.mypy_cache/
+.ruff_cache/