--- a/python/.gitignore
+++ b/python/.gitignore
@@ -28,6 +28,7 @@ pip-log.txt
 
 # Unit test / coverage reports
 .coverage
+coverage.xml
 .tox
 nosetests.xml