--- a/python/pyproject.toml
+++ b/python/pyproject.toml
@@ -131,5 +131,10 @@ python =
     pypy-3.10: pypy310
 """
 
+[tool.ruff.lint]
+select = [
+  "UP",
+]
+
 [tool.pytest.ini_options]
 testpaths = ["tests"]