--- a/python/pyproject.toml
+++ b/python/pyproject.toml
@@ -57,11 +57,6 @@ line-length = 120
 target-version = ["py39", "py310", "py311", "py312", "py313"]
 verbose = true
 
-[tool.isort]
-line_length = 120
-multi_line_output = 3
-profile = "black"
-
 [tool.mypy]
 files = "src/**/*.py"
 install_types = true
@@ -74,6 +69,21 @@ warn_unused_configs = true
 [[tool.mypy.overrides]]
 ignore_missing_imports = true
 module = [
+]
+
+# Redundant. Automatically introduced by `pretty-format-toml` pre-commit job
+[tool.pytest]
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+
+# Redundant. Automatically introduced by `pretty-format-toml` pre-commit job
+[tool.ruff]
+
+[tool.ruff.lint]
+select = [
+  "I",
+  "UP"
 ]
 
 # Once https://github.com/tox-dev/tox/issues/999 is released and available, migrate to the new tox approach
@@ -130,11 +140,3 @@ python =
     pypy-3.9: pypy39
     pypy-3.10: pypy310
 """
-
-[tool.ruff.lint]
-select = [
-  "UP",
-]
-
-[tool.pytest.ini_options]
-testpaths = ["tests"]