--- a/python/pyproject.toml
+++ b/python/pyproject.toml
@@ -1,6 +1,6 @@
 [build-system]
-build-backend = "setuptools.build_meta"
-requires = ["setuptools>=61.0", "wheel"]
+requires = ["uv_build>=0.8.4,<0.9.0"]
+build-backend = "uv_build"
 
 [project]
 authors = [
@@ -24,12 +24,16 @@ classifiers = [
 ]
 dependencies = []
 description = "Cucumber Messages is a message protocol for representing results and other information from Cucumber. "
-license = {text = "MIT"}
+keywords = ["gherkin", "cucumber", "bdd"]
+license = "MIT"
+license-files = [
+  "LICENSE",
+]
 maintainers = [
   {name = "Konstantin Goloveshko", email = "kostya.goloveshko@gmail.com"}
 ]
 name = "cucumber-messages"
-readme = {file = "README.md", content-type = "text/markdown"}
+readme = "README.md"
 requires-python = ">=3.9"
 urls = {Repository = "https://github.com/cucumber/messages"}
 version = "28.1.0"