--- a/python/pyproject.toml
+++ b/python/pyproject.toml
@@ -1,6 +1,6 @@
 [build-system]
-requires = ["uv_build>=0.8.4,<0.9.0"]
 build-backend = "uv_build"
+requires = ["uv_build>=0.8.4,<0.9.0"]
 
 [project]
 authors = [
@@ -27,7 +27,7 @@ description = "Cucumber Messages is a message protocol for representing results 
 keywords = ["gherkin", "cucumber", "bdd"]
 license = "MIT"
 license-files = [
-  "LICENSE",
+  "LICENSE"
 ]
 maintainers = [
   {name = "Konstantin Goloveshko", email = "kostya.goloveshko@gmail.com"}
@@ -65,14 +65,7 @@ warn_unused_configs = true
 
 [[tool.mypy.overrides]]
 ignore_missing_imports = true
-module = [
-]
-
-# Redundant. Automatically introduced by `pretty-format-toml` pre-commit job
-[tool.pytest]
-
-[tool.pytest.ini_options]
-testpaths = ["tests"]
+module = []
 
 [tool.ruff]
 line-length = 120