--- a/python/pyproject.toml
+++ b/python/pyproject.toml
@@ -62,10 +62,6 @@ show_error_codes = true
 warn_return_any = true
 warn_unused_configs = true
 
-[[tool.mypy.overrides]]
-ignore_missing_imports = true
-module = []
-
 [tool.ruff]
 line-length = 120