--- a/python/pyproject.toml
+++ b/python/pyproject.toml
@@ -80,7 +80,7 @@ extend-select = [
 "src/cucumber_messages/__init__.py" = ["F403"]
 
 [tool.tox]
-env_list = ["lint", "3.13", "3.12", "3.11", "3.10", "3.9", "pypy310", "pypy39"]
+env_list = ["lint", "3.13", "3.12", "3.11", "3.10", "3.9", "pypy311", "pypy310"]
 requires = ["tox>=4.22"]
 
 [tool.tox.env.lint]