### 🤔 What's changed?

- [uv](https://docs.astral.sh/uv/) package management (and cache) - over [pip](https://pip.pypa.io/en/stable/), [pre-commit](https://pre-commit.com) and [tox](https://tox.wiki/en/4.30.2/)
- [uv](https://docs.astral.sh/uv/) Python management (and cache) - over [setup-python](https://github.com/actions/setup-python), [virtualenv](https://virtualenv.pypa.io/en/latest/) (via [tox](https://tox.wiki/en/4.30.2/)) and user installed Python distributions
- [uv build backend](https://docs.astral.sh/uv/concepts/build-backend/) over [setuptools](https://setuptools.pypa.io/en/latest/build_meta.html)
- [ruff](https://docs.astral.sh/ruff/) linter and formatter over [black](https://black.readthedocs.io/en/stable/), [isort](https://pycqa.github.io/isort/) and [pyugrade](https://github.com/asottile/pyupgrade)
- `pyproject.toml` configuration over both `pyproject.toml` and `pytest.ini`
  - [tox native toml configuration](https://tox.wiki/en/stable/config.html#pyproject-toml-native) over [legacy ini embedded within toml](https://tox.wiki/en/stable/config.html#pyproject-toml-ini)
- GitHub Actions matrix (see [flake8 workflow](https://github.com/PyCQA/flake8/blob/main/.github/workflows/main.yml)) over [tox-gh-actions](https://pypi.org/project/tox-gh-actions/) Python dependency
- [GitHub Codecov Action](https://github.com/codecov/codecov-action) over deprecated [codecov](https://github.com/codecov/codecov-python) Python dependency
- [pre-commit](https://pre-commit.com) linting trigger over both [pre-commit](https://pre-commit.com) and [tox](https://tox.wiki/en/4.30.2/) (triggering [mypy](https://mypy.readthedocs.io/en/stable/))
  - Trigger linting once in workflow over separately running type checking against all Pythons
- [Dependency groups](https://peps.python.org/pep-0735/) (PEP735) dependency declaration over [optional dependencies](https://peps.python.org/pep-0631/#optional-dependencies) (PEP631), [tox](https://tox.wiki/en/4.30.2/), and GitHub workflows 
- [pypy311](https://pypy.org/posts/2025/01/towards-pypy311-an-update.html) testing over end of life [pypy39](https://pypy.org/posts/2024/08/pypy-v7317-release.html#pypy-versions-and-speed-pypy-org)
- [Legacy .egg-info metadata](https://packaging.python.org/en/latest/discussions/package-formats/#what-about-eggs) and tests removed from distribution

### ⚡️ What's your motivation? 

- Improve CI performance

  | Metric | [Previous](https://github.com/cucumber/messages/actions/runs/17550938463) | [Now](https://github.com/cucumber/messages/actions/runs/17710887050) | Delta | % Change |
  |--------|----------|-----|-------|----------|
  | **Total Runner Time** | 23m 55s<br>(1435s) | 16m 47s<br>(1007s) | -7m 8s<br>(-428s) | -29.8% |
  | **Total Python Test Runner Time** | [9m 49s<br>(589s)](https://github.com/cucumber/messages/actions/runs/17550938463/usage) | [2m 58s<br>(178s)](https://github.com/cucumber/messages/actions/runs/17710887050/usage) | -6m 51s<br>(-411s) | -69.8% |
  | **Python Test Workflow Duration** | [1m 59s<br>(119s)](https://github.com/cucumber/messages/actions/runs/17550938463) | [40s](https://github.com/cucumber/messages/actions/runs/17710887050) | -1m 19s<br>(-79s) | -66.4% |
  | **Python Test Fastest Job** | [47s](https://github.com/cucumber/messages/actions/runs/17550938463/job/49843006379) | [12s](https://github.com/cucumber/messages/actions/runs/17710887050/job/50329404794?pr=324) | 35s | -74.5% |
  | **Python Test Average Job** | 1m 6s<br>(66s) | 19.8s | -46.2s | -70% |
  | **Python Codecov Step** | 213s<br>(avg. 23s) | 23s<br>(avg. 2.6s) | -190s<br>(-3m 10s) | -89.2% |

  - 69.8% reduction in runner uptime - saving $87.29 and 5949 runner minutes compared to [6 days total runner usage for this workflow over the past year](https://github.com/cucumber/messages/actions/metrics/usage?dateRangeType=DATE_RANGE_TYPE_LAST_YEAR) (assuming [GitHub Actions pricing](https://docs.github.com/en/billing/concepts/product-billing/github-actions#minute-multipliers))
  - 66.4% faster workflow - saving 0.9 days where the workflow would otherwise be running, across a thousand runs
  - 70% faster average job execution - continuous Python feedback in 12s

- Improve maintainability
  - Prefer tooling and language agnostic, portable solutions (GitHub Actions) over alternatives e.g. [codecov](https://github.com/codecov/codecov-python) and [tox-gh-actions](https://pypi.org/project/tox-gh-actions/) Python dependencies; eliminating workflow prerequisites and having greater familiarity across polyglot maintainers
  - Streamline supply chain e.g. eliminate [codecov](https://github.com/codecov/codecov-python) and [tox-gh-actions](https://pypi.org/project/tox-gh-actions/), one linter and formatter ([ruff](https://docs.astral.sh/ruff/)) instead of three ([black](https://black.readthedocs.io/en/stable/), [isort](https://pycqa.github.io/isort/), [pyugrade](https://github.com/asottile/pyupgrade)), etc.
    - Eliminate duplicate dependency references such as [mypy](https://mypy.readthedocs.io/en/stable/) across `pyproject.toml`, [pre-commit](https://pre-commit.com) and [tox](https://tox.wiki/en/4.30.2/)
    - Freeze [pre-commit](https://pre-commit.com) dependencies and use [GitHub Codecov Action](https://github.com/codecov/codecov-action) for [integrity checking with SHASUM verification](https://docs.codecov.com/docs/codecov-uploader#using-the-cli-to-upload-reports-with-codecovio-cloud) for increased security
  - Standardise - _"There should be one-- and preferably only one --obvious way to do it."_ ([PEP20](https://peps.python.org/pep-0020/)) - for consistent development and maintenance across Cucumber, emulating cucumber/gherkin#439
    - Coding standard parity with [gherkin](https://github.com/cucumber/gherkin/blob/06fc2deaeda590b8dcd1c23562471dfd3c714816/python/pyproject.toml#L82-L93)
  - Remove legacy references requiring comprehension to overhaul tooling e.g. removing the [deprecated `distshare` tox configuration](https://tox.wiki/en/3.1.3/config.html#tox-global-settings) that reapplies default
- Improve developer experience
  - Fixes [mypy](https://mypy.readthedocs.io/en/stable/) cache requiring deletion before subsequent [tox](https://tox.wiki/en/stable/index.html) runs
  - Fixes [tox](https://tox.wiki/en/stable/index.html) running tests against all Python versions with linux only - having been configured for GitHub Actions using [tox-gh-actions](https://pypi.org/project/tox-gh-actions/) rather than local development

### 🏷️ What kind of change is this?

- :bank: Refactoring/debt/DX (improvement to code design, tooling, etc. without changing behaviour)

### ♻️ Anything particular you want feedback on?

- Codecov GitHub Action not uploading with _"Token required because branch is protected"_ error ([#](https://github.com/cucumber/messages/actions/runs/17710887050/job/50329404794#step:5:272))... which historically has been an intermittent issue (see codecov/codecov-action#1709); hitting a [429 rate limit](https://github.com/cucumber/messages/actions/runs/17550938463/job/49843006354#step:6:46) with a comparable error message _"Please upload with the Codecov repository upload token to resolve issue"_ before the change... though the [same configuration](https://github.com/kieran-ryan/pyprojectsort/blob/a2666892e26965b4e162ac406dcbfe5c50ea4693/.github/workflows/main.yml#L25-L29) is [running on pyprojectsort](https://github.com/kieran-ryan/pyprojectsort/pull/83). Assumption that the issue is pre-existing and that the change in error message is in consequence of migrating from the deprecated CLI to the GitHub Action. Under those circumstances, would suggest to tackle separately if not immediately resolvable through a token change or other means - with no material impact.
- Would advocate for [pre-commit.ci](https://pre-commit.ci) integration to autoupdate dependencies and improve linting performance; the caveat being it would trigger for a negligible ~2s on each run to check whether there are changed files to lint - compared to a GitHub workflow which would only trigger when there are Python changes. Increasingly desirable with wider adoption of polyglot linting hooks over time.
- Having optimised the slowest, most expensive workflow, assessing whether the remaining bottlenecks to rapid CI feedback (cpp and javascript workflows) could be optimised, would be a worthwhile endeavour.
- Windows runner performance appears inconsistent on dependency download. Further caching optimisation across uv, tox and pre-commit may mitigate.

### 📋 Checklist:

- [x] I agree to respect and uphold the [Cucumber Community Code of Conduct](https://github.com/cucumber/.github/tree/main?tab=coc-ov-file)
- [x] Users should know about my change
  - [x] I have added an entry to the "Unreleased" section of the [**CHANGELOG**](../blob/main/CHANGELOG.md), linking to this pull request.