@mpkorstanje 2025-09-14
Cheers! Again, much appreciated!

> [Legacy .egg-info metadata](https://packaging.python.org/en/latest/discussions/package-formats/#what-about-eggs) and tests removed from distribution

Is this a breaking change? If so, please put it in the `CHANGELOG`?

> Codecov GitHub Action not uploading with "Token required because branch is protected" error

I'll have a look. 

> Would advocate for [pre-commit.ci](https://pre-commit.ci/) integration to autoupdate dependencies and improve linting performance; the caveat being it would trigger for a negligible ~2s on each run to check whether there are changed files to lint - compared to a GitHub workflow which would only trigger when there are Python changes. Increasingly desirable with wider adoption of polyglot linting hooks over time.

Not opposed in principle.

The biggest limitation to using any sort of pre-commit is that the tools triggered by pre-commit must be installed on a developers machine. This is a problem in a polyglot repo because not everyone will have all tools installed. At a glance it looks like pre-commit.ci avoids that problem.

Updating dependencies should preferably be done through Renovate.